+++ src/domain/index.ts
@@ -1,30 +1,7 @@
-/**
- * Domain layer index file
- * Exports all domain functionality to be used by adapters
- */
-
-// Tasks domain exports
-export { 
-  listTasksFromParams, 
-  getTaskFromParams,
-  getTaskStatusFromParams,
-  setTaskStatusFromParams 
-} from "./tasks.js";
-
-// Session domain exports
-export { 
-  getSessionFromParams,
-  listSessionsFromParams,
-  startSessionFromParams,
-  deleteSessionFromParams,
-  getSessionDirFromParams,
-  updateSessionFromParams
-} from "./session.js";
-
-// Git domain exports
-export {
-  createPullRequestFromParams,
-  commitChangesFromParams
-} from "./git.js";
-
-// Add other domain exports as they are implemented 
+export * from "./git.js";
+export * from "./session.js";
+export * from "./tasks.js";
+export * from "./rules.js";
+export * from "./repo-utils.js";
+export * from "./workspace.js";
+export * from "./init.js";